+++ src/compile/mark/encode/conditional.ts
@@ -7,21 +7,37 @@ import {parseSelectionPredicate} from '../../selection/parse';
 import {UnitModel} from '../../unit';
 
 /**
- * Return a mixin that includes a Vega production rule for a Vega-Lite conditional channel definition
- * or a simple mixin if channel def has no condition.
+ * Return a VgEncodeEntry that includes a Vega production rule for a scale channel's encoding or guide encoding, which includes:
+ * (1) the conditional rules (if provided as part of channelDef)
+ * (2) invalidValueRef for handling invalid values (if provided as a parameter of this method)
+ * (3) main reference for the encoded data.
  */
-export function wrapCondition<CD extends ChannelDef | GuideEncodingConditionalValueDef>(
-  model: UnitModel,
-  channelDef: CD,
-  vgChannel: string,
-  refFn: (cDef: CD) => VgValueRef
-): VgEncodeEntry {
+export function wrapCondition<CD extends ChannelDef | GuideEncodingConditionalValueDef>({
+  model,
+  channelDef,
+  vgChannel,
+  invalidValueRef,
+  mainRefFn
+}: {
+  model: UnitModel;
+  channelDef: CD;
+  vgChannel: string;
+
+  /**
+   * invalidValue for a scale channel if the invalidDataMode is include for the channel.
+   * For scale channel with other invalidDataMode or non-scale channel, this value should be undefined.
+   */
+  invalidValueRef: VgValueRef | undefined;
+  mainRefFn: (cDef: CD) => VgValueRef;
+}): VgEncodeEntry {
   const condition = isConditionalDef<CD>(channelDef) && channelDef.condition;
-  const valueRef = refFn(channelDef);
+
+  let valueRefs: VgValueRef[] = [];
+
   if (condition) {
     const conditions = array(condition);
-    const vgConditions = conditions.map(c => {
-      const conditionValueRef = refFn(c);
+    valueRefs = conditions.map(c => {
+      const conditionValueRef = mainRefFn(c);
       if (isConditionalParameter<any>(c)) {
         const {param, empty} = c;
         const test = parseSelectionPredicate(model, {param, empty});
@@ -31,10 +47,21 @@ export function wrapCondition<CD extends ChannelDef | GuideEncodingConditionalVa
         return {test, ...conditionValueRef};
       }
     });
-    return {
-      [vgChannel]: [...vgConditions, ...(valueRef !== undefined ? [valueRef] : [])]
-    };
-  } else {
-    return valueRef !== undefined ? {[vgChannel]: valueRef} : {};
   }
+
+  if (invalidValueRef !== undefined) {
+    valueRefs.push(invalidValueRef);
+  }
+
+  const mainValueRef = mainRefFn(channelDef);
+  if (mainValueRef !== undefined) {
+    valueRefs.push(mainValueRef);
+  }
+
+  if (valueRefs.length > 1) {
+    return {[vgChannel]: valueRefs};
+  } else if (valueRefs.length === 1) {
+    return {[vgChannel]: valueRefs[0]};
+  }
+  return {};
 }