+++ src/scale.ts
@@ -8,12 +8,13 @@ import {
   TimeInterval,
   TimeIntervalStep
 } from 'vega';
-import {isString} from 'vega-util';
 import type {ColorScheme} from 'vega-typings';
+import {isString} from 'vega-util';
 import * as CHANNEL from './channel';
 import {Channel, isColorChannel} from './channel';
 import {DateTime} from './datetime';
 import {ExprRef} from './expr';
+import {ScaleInvalidDataConfigMixins} from './invalid';
 import * as log from './log';
 import {ParameterExtent} from './selection';
 import {NOMINAL, ORDINAL, QUANTITATIVE, TEMPORAL, Type} from './type';
@@ -180,7 +181,7 @@ export function isContinuousToDiscrete(type: ScaleType): type is 'quantile' | 'q
   return CONTINUOUS_TO_DISCRETE_SCALES.has(type);
 }
 
-export interface ScaleConfig<ES extends ExprRef | SignalRef> {
+export interface ScaleConfig<ES extends ExprRef | SignalRef> extends ScaleInvalidDataConfigMixins {
   /**
    * If true, rounds numeric output values to integers.
    * This can be helpful for snapping to the pixel grid.
@@ -315,7 +316,7 @@ export interface ScaleConfig<ES extends ExprRef | SignalRef> {
   maxBandSize?: number;
 
   /**
-   * The default min value for mapping quantitative fields to bar and tick's size/bandSize scale with zero=false.
+   * The default min value for mapping quantitative fields to bar and tick's size/bandSize scale.
    *
    * __Default value:__ `2`
    *
@@ -324,7 +325,7 @@ export interface ScaleConfig<ES extends ExprRef | SignalRef> {
   minBandSize?: number;
 
   /**
-   * The default max value for mapping quantitative fields to text's size/fontSize.
+   * The default max value for mapping quantitative fields to text's size/fontSize scale.
    *
    * __Default value:__ `40`
    *
@@ -333,7 +334,7 @@ export interface ScaleConfig<ES extends ExprRef | SignalRef> {
   maxFontSize?: number;
 
   /**
-   * The default min value for mapping quantitative fields to tick's size/fontSize scale with zero=false
+   * The default min value for mapping quantitative fields to text's size/fontSize scale.
    *
    * __Default value:__ `8`
    *
@@ -362,7 +363,7 @@ export interface ScaleConfig<ES extends ExprRef | SignalRef> {
   maxOpacity?: number;
 
   /**
-   * Default minimum value for point size scale with zero=false.
+   * Default minimum value for point size scale.
    *
    * __Default value:__ `9`
    *
@@ -377,7 +378,7 @@ export interface ScaleConfig<ES extends ExprRef | SignalRef> {
   maxSize?: number;
 
   /**
-   * Default minimum strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks with zero=false.
+   * Default minimum strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks.
    *
    * __Default value:__ `1`
    *
@@ -443,7 +444,7 @@ export const defaultScaleConfig: ScaleConfig<SignalRef> = {
   maxOpacity: 0.8,
 
   // FIXME: revise if these *can* become ratios of width/height step
-  minSize: 9, // Point size is area. For square point, 9 = 3 pixel ^ 2, not too small!
+  minSize: 4, // Point size is area. For square point, 9 = 3 pixel ^ 2, not too small!
 
   minStrokeWidth: 1,
   maxStrokeWidth: 4,
@@ -481,7 +482,7 @@ export type Domain =
 
 export type Scheme = string | SchemeParams;
 
-export function isExtendedScheme(scheme: Scheme | SignalRef | string[]): scheme is SchemeParams {
+export function isExtendedScheme(scheme: Scheme | SignalRef): scheme is SchemeParams {
   return !isString(scheme) && !!scheme['name'];
 }
 
@@ -609,9 +610,11 @@ export interface Scale<ES extends ExprRef | SignalRef = ExprRef | SignalRef> {
    *
    * Discrete color schemes may be used with [discrete](https://vega.github.io/vega-lite/docs/scale.html#discrete) or [discretizing](https://vega.github.io/vega-lite/docs/scale.html#discretizing) scales. Continuous color schemes are intended for use with color scales.
    *
+   * To set a custom scheme, instead set the list of values [as the scale range](https://vega.github.io/vega-lite/docs/scale.html#2-setting-the-range-property-to-an-array-of-valid-css-color-strings).
+   *
    * For the full list of supported schemes, please refer to the [Vega Scheme](https://vega.github.io/vega/docs/schemes/#reference) reference.
    */
-  scheme?: ColorScheme | string[] | SchemeParams | ES;
+  scheme?: ColorScheme | SchemeParams | ES;
 
   /**
    * The alignment of the steps within the scale range.